requirements.txt
Flask==2.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Werkzeug==2.0.1
urllib3==1.22
pytest==7.4.0
puzpy==0.2.5